Refactor Feedback to its own component.
